third_party/blink/web_tests/platform/generic/http/tests/persistent-origin-trials/window-origin-trial.https-expected.txt

This is a testharness.js-based test.
PASS TestPrimaryPageHasNoTrial
FAIL PersistentInChild assert_equals: expected "FrobulatePersistent" but got ""
Harness: the test ran to completion.
